Render production USD on the hdSilk managed backends
hdSilk could not draw real production assets: skinned meshes threw
"triangle index is outside the point array", point-instanced assets threw
"Creating Instancer not supported", a single bad prim aborted the whole
frame, and meshes above 65,535 vertices were silently truncated.

Add point instancing via a real HdSilkInstancer, and register
extComputation as a supported Sprim type so Hydra actually creates the
computation UsdSkel depends on; without it, pulling computed primvars for
a skinned mesh faults. Skinned points are now read from computed primvars
and refreshed whenever topology refreshes.

Page ABI moves from 2 to 3. Instance fields become meaningful: a prototype
publishes one record per resolved instance with its own transform, so
retained meshes are keyed by (path, instance index) and a removal retires
exactly one identity. Serialization now isolates failures per prim,
skipping and counting invalid records instead of aborting the page.
Indices are 32-bit end to end across the wire, retained state, and the
D3D12, Vulkan, and Metal backends.

eng/run-viewer.ps1 gains -StageAssetRoot; staging only the root layer left
every reference unresolved, which silently produced an empty stage.

Verified on Windows: native CTest 7/7, full managed suite 15/15 projects
including 27/27 package tests, RHI probe, 42 checked shader artifacts,
format, line length, documentation, and actionlint. End to end the
instanced chess set renders 21 draws, the skinned bike 331, and the
OpenPBR shader playground 341, with no skipped meshes.

Copilot-Session: 8e6402f2-8247-4692-882f-4757998bc0a2

+++ src/OpenUsd.Rendering.Silk.Metal/MetalSilkGraphicsDevice.Offscreen.cs
@@ -635,7 +635,7 @@ public sealed partial class MetalSilkGraphicsDevice
                                 encoder.DrawIndexedPrimitives(
                                     MTLPrimitiveType.Triangle,
                                     command.IndexCount,
-                                    MTLIndexType.UInt16,
+                                    MTLIndexType.UInt32,
                                     indexBuffer.Buffer,
                                     0);
                                 encoder.EndEncoding();
