+++ src/OpenUsd.Rendering.Silk.Vulkan/VulkanSilkGraphicsDevice.Offscreen.cs
@@ -600,7 +600,7 @@ public sealed unsafe partial class VulkanSilkGraphicsDevice
                             nativeCommands,
                             indexBuffer.Buffer,
                             0,
-                            IndexType.Uint16);
+                            IndexType.Uint32);
                         DescriptorSet descriptorSet = drawResource.DescriptorSet;
                         _api.CmdBindDescriptorSets(
                             nativeCommands,
